requirements.txt
asgiref==3.7.2
black==23.3.0
cffi==1.15.1
click==8.1.4
cryptography==41.0.1
Django==4.2.3
django-admin-volt==1.0.10
django-cors-headers==4.1.0
django-passkeys==1.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
fido2==1.1.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
ua-parser==0.16.1
user-agents==2.2.0
